날씨 관련 추가 및 context 수정 - 12/23 강민
