+++ src/main/moivo_react/src/App.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { AuthProvider } from './contexts/AuthContext';
 import Main_index from './components/Main_index';
+import MainProvider from './contexts/MainContext';
 import User_login from './containers/user/user_login';
 import User_signup from './containers/user/user_signup';
 import ProductBoard from './containers/product/product_board'; 
@@ -74,13 +75,15 @@ const routeConfig = [
 const App = () => {
   return (
     <Router>
-        <AuthProvider>
-                <Routes>
-                {routeConfig.map(({ path, element }, index) => (
-                  <Route key={index} path={path} element={element} />
-                ))}
-              </Routes>
-        </AuthProvider>
+      <AuthProvider>
+        <MainProvider>
+          <Routes>
+            {routeConfig.map(({ path, element }, index) => (
+              <Route key={index} path={path} element={element} />
+            ))}
+          </Routes>
+        </MainProvider>
+      </AuthProvider>
     </Router>
   );
 };
