+++ src/main/moivo_react/src/contexts/MainContext.jsx
@@ -1,10 +1,24 @@
-import React from 'react';
-import { createContext, useContext, useEffect } from 'react';
+import React, { createContext, useContext, useEffect, useState, useRef } from 'react';
+import AOS from 'aos';
+import 'aos/dist/aos.css';
 
+// Context 생성 및 export
 const MainContext = createContext();
 export const useMainIndex = () => useContext(MainContext);
 
-const MainProvider = ({children}) => {
+const MainProvider = ({ children }) => {
+    const [animate, setAnimate] = useState(true);
+    const [menuOpen, setMenuOpen] = useState(false);
+    const parentDivRef = useRef(null);
+    const childDivRef = useRef(null);
+
+    // ES6 import 방식으로 이미지 가져오기
+    const slides = [
+        { src: new URL('../assets/image/1_outer.jpg', import.meta.url).href },
+        { src: new URL('../assets/image/2_outer.jpg', import.meta.url).href },
+        { src: new URL('../assets/image/4_outer.jpg', import.meta.url).href },
+        { src: new URL('../assets/image/6_outer.jpg', import.meta.url).href },
+    ];
 
     useEffect(() => {
         const link = document.createElement('link');
@@ -12,14 +26,66 @@ const MainProvider = ({children}) => {
         link.rel = "stylesheet";
         document.head.appendChild(link);
 
-        // 컴포넌트가 언마운트될 때 link 태그 제거
         return () => {
             document.head.removeChild(link);
         };
     }, []);
 
+    useEffect(() => {
+        AOS.init({
+            duration: 2000,
+            once: false,
+            offset: 0,
+        });
+        AOS.refresh();
+    }, []);
+
+    useEffect(() => {
+        const handleScroll = () => {
+            const parentDiv = parentDivRef.current;
+            const viewportHeight = window.innerHeight;
+            const fromViewportToParentHeight = parentDiv.getBoundingClientRect().top;
+            const scrollPassedAmount = viewportHeight - fromViewportToParentHeight;
+
+            let divHeight = parentDiv.clientHeight > viewportHeight ? viewportHeight : parentDiv.clientHeight;
+            let scrollRate = (scrollPassedAmount / divHeight) * 100;
+
+            if (scrollRate < 0) {
+                scrollRate = 0;
+            } else if (scrollRate > 100) {
+                scrollRate = 100;
+            }
+
+            const childDiv = childDivRef.current;
+            childDiv.style.transform = `scale(${scrollRate / 100})`;
+        };
+
+        window.addEventListener('scroll', handleScroll);
+
+        return () => {
+            window.removeEventListener('scroll', handleScroll);
+        };
+    }, []);
+
+    const onStop = () => setAnimate(false);
+    const onRun = () => setAnimate(true);
+
+    const toggleMenu = () => setMenuOpen((prev) => !prev);
+
     return (
-        <MainContext.Provider value=''>
+        <MainContext.Provider
+            value={{
+                animate,
+                setAnimate,
+                onStop,
+                onRun,
+                menuOpen,
+                toggleMenu,
+                slides,
+                parentDivRef,
+                childDivRef,
+            }}
+        >
             {children}
         </MainContext.Provider>
     );
